lan78xx: New simple patch to prevent some crashes
Roll back of previous patch and backport of raspberrypi/linux commit
661230038a8736ba9023978c53cd21cd65739406

Alexander Graf submitted a patch to solve a problem when connecting to
the device before the interface is brought up, but with that patch
applied the interface fails to work (for me and others). This simpler
patch should be considered a stepping stone that may prevent some of the
crashes seen when the interface is reset but without solving the early
access problem.

See:  #2449